backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
typer==0.12.3
pyyaml==6.0.1
neo4j==5.20.0
tenacity==8.2.3
kubernetes==28.1.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
httpx==0.27.0
loguru==0.7.2
pytest==8.2.1
pytest-asyncio==0.23.7
respx==0.21.1
networkx==3.3
matplotlib==3.8.4
